fix(test): Moved CpanCliDetectableTest to functional test folder, deleted tests that are redundant
